backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
psycopg2-binary==2.9.9
requests==2.31.0
boto3==1.34.0  # For S3
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
google-auth==2.27.0
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
